refactor: restructure and clean up utility functions and tests
