+++ src/module.ts
@@ -1,4 +1,5 @@
-import { addComponent, addImports, addImportsSources, addPlugin, addTemplate, addTypeTemplate, defineNuxtModule, resolveModule } from '@nuxt/kit'
+import { addComponent, addImports, addImportsSources, addPlugin, addTemplate, addTypeTemplate, addVitePlugin, defineNuxtModule, resolveModule, useNitro } from '@nuxt/kit'
+import { defu } from 'defu'
 import { setupPages } from './pages'
 import { setupNitro } from './nitro'
 import { extendBundler } from './bundler'
@@ -8,15 +9,13 @@ import type { I18nPublicRuntimeConfig, LocaleObject, NuxtI18nOptions } from './t
 import type { Locale } from 'vue-i18n'
 import { createContext } from './context'
 import { prepareOptions } from './prepare/options'
-import { resolveLocaleInfo } from './prepare/locale-info'
-import { prepareHMR } from './prepare/runtime'
-import { prepareRuntimeConfig } from './prepare/runtime-config'
-import { prepareBuildManifest } from './prepare/build-manifest'
-import { prepareStrategy } from './prepare/strategy'
 import { prepareTypeGeneration } from './prepare/type-generation'
 import { relative } from 'pathe'
 import { generateTemplateNuxtI18nOptions } from './template'
-import { generateI18nTypes, generateLoaderOptions } from './gen'
+import { generateI18nTypes, generateLoaderOptions, simplifyLocaleOptions } from './gen'
+import { applyLayerOptions, resolveLayerVueI18nConfigInfo } from './layers'
+import { filterLocales, resolveLocales } from './utils'
+import { isString } from '@intlify/shared'
 
 export * from './types'
 
@@ -35,14 +34,8 @@ export default defineNuxtModule<NuxtI18nOptions>({
   async setup(i18nOptions, nuxt) {
     const ctx = createContext(i18nOptions, nuxt)
 
-    /**
-     * prepare options
-     */
     prepareOptions(ctx, nuxt)
 
-    /**
-     * auto imports
-     */
     addComponent({
       name: 'NuxtLinkLocale',
       filePath: ctx.resolver.resolve(ctx.runtimeDir, 'components/NuxtLinkLocale'),
@@ -86,9 +79,6 @@ export default defineNuxtModule<NuxtI18nOptions>({
       '@intlify/message-compiler',
     ]
 
-    /**
-     * alias and transpile dependencies
-     */
     for (const dep of deps) {
       if (dep === 'vue-i18n' || dep === '@intlify/core') { continue }
       nuxt.options.alias[dep] = resolveModule(dep)
@@ -98,9 +88,6 @@ export default defineNuxtModule<NuxtI18nOptions>({
     nuxt.options.alias['@intlify/core'] = resolveModule(`@intlify/core/dist/core.node`)
     nuxt.options.build.transpile.push('@nuxtjs/i18n', ...deps)
 
-    /**
-     * alias and transpile runtime and internals
-     */
     nuxt.options.alias['#i18n'] = ctx.resolver.resolve('./runtime/composables/index')
     nuxt.options.alias['#i18n-kit'] = ctx.resolver.resolve('./runtime/kit')
     nuxt.options.alias['#internal-i18n-types'] = ctx.resolver.resolve('./types')
@@ -115,7 +102,7 @@ export default defineNuxtModule<NuxtI18nOptions>({
     nuxt.options.vite.optimizeDeps.exclude.push(...deps)
 
     /**
-     * typescript hoist dependencies and include i18n directories
+     * hoist deps and include i18n directories
      */
     nuxt.options.typescript.hoist ||= []
     nuxt.options.typescript.hoist.push(...deps)
@@ -125,9 +112,6 @@ export default defineNuxtModule<NuxtI18nOptions>({
       ...ctx.i18nLayers.map(l => relative(nuxt.options.buildDir, l.i18nDir + '/**/*')),
     )
 
-    /**
-     * add plugin and templates
-     */
     addPlugin(ctx.resolver.resolve('./runtime/plugins/i18n'))
     if (nuxt.options.dev || nuxt.options._prepare) {
       addPlugin(ctx.resolver.resolve('./runtime/plugins/dev'))
@@ -151,7 +135,27 @@ export default defineNuxtModule<NuxtI18nOptions>({
       getContents: () => generateI18nTypes(nuxt, ctx),
     })
 
-    prepareHMR(ctx, nuxt)
+    /**
+     * HMR plugin
+     */
+    if (nuxt.options.dev && ctx.options.hmr) {
+      addVitePlugin({
+        name: 'i18n:options-hmr',
+        configureServer(server) {
+          const reloadClient = () => server.ws.send({ type: 'full-reload' })
+
+          server.ws.on('i18n:options-complex-invalidation', () => {
+            // await dev reload if type generation is enabled
+            if (ctx.options.experimental.typedOptionsAndMessages) {
+              useNitro().hooks.hookOnce('dev:reload', reloadClient)
+              return
+            }
+
+            reloadClient()
+          })
+        },
+      })
+    }
 
     /**
      * generate vue-i18n and messages types using runtime server endpoint
@@ -159,42 +163,82 @@ export default defineNuxtModule<NuxtI18nOptions>({
     await prepareTypeGeneration(ctx, nuxt)
 
     /**
-     * allow other modules to register i18n hooks, then merge locales
+     * allow other modules to register i18n hooks - locales and options will be resolved in this hook
      */
     nuxt.hook('modules:done', async () => {
-      /**
-       * resolve locale info and vue-i18n config path
-       */
-      await resolveLocaleInfo(ctx, nuxt)
+      ctx.options.locales = await applyLayerOptions(ctx, nuxt)
+      ctx.options.locales = filterLocales(ctx, nuxt)
+
+      ctx.normalizedLocales = ctx.options.locales.map(x => (isString(x) ? { code: x, language: x } : x))
+      ctx.localeCodes = ctx.normalizedLocales.map(locale => locale.code)
+      ctx.localeInfo = resolveLocales(nuxt.options.srcDir, ctx.normalizedLocales, nuxt.vfs)
+
+      ctx.vueI18nConfigPaths = await resolveLayerVueI18nConfigInfo(ctx.options)
 
       /**
-       * setup runtime config
+       * expose i18n options via runtime config for use in app/server contexts
        */
-      prepareRuntimeConfig(ctx, nuxt)
+      // @ts-expect-error generated type
+      nuxt.options.runtimeConfig.public.i18n = defu(nuxt.options.runtimeConfig.public.i18n, {
+        baseUrl: ctx.options.baseUrl,
+        defaultLocale: ctx.options.defaultLocale,
+        rootRedirect: ctx.options.rootRedirect,
+        redirectStatusCode: ctx.options.redirectStatusCode,
+        skipSettingLocaleOnNavigate: ctx.options.skipSettingLocaleOnNavigate,
+        locales: ctx.options.locales,
+        detectBrowserLanguage: ctx.options.detectBrowserLanguage ?? DEFAULT_OPTIONS.detectBrowserLanguage,
+        experimental: ctx.options.experimental,
+        domainLocales: Object.fromEntries(
+          ctx.options.locales.map((l) => {
+            if (typeof l === 'string') {
+              return [l, { domain: '' }]
+            }
+            return [l.code, { domain: l.domain ?? '' }]
+          }),
+        ) as I18nPublicRuntimeConfig['domainLocales'],
+      })
 
-      /**
-       * setup nuxt pages
-       */
-      await setupPages(ctx, nuxt)
+      nuxt.options.runtimeConfig.public.i18n.locales = simplifyLocaleOptions(ctx, nuxt)
 
       /**
        * ignore `/` during prerender when using prefixed routing
        */
-      prepareStrategy(ctx, nuxt)
+      if (ctx.options.strategy === 'prefix' && nuxt.options.nitro.static) {
+        const localizedEntryPages = ctx.localeCodes.map(x => '/' + x)
+        nuxt.hook('nitro:config', (config) => {
+          config.prerender ??= {}
+
+          // ignore `/` which is added by nitro by default
+          config.prerender.ignore ??= []
+          config.prerender.ignore.push(/^\/$/)
+
+          // add localized routes as entry pages for prerendering
+          config.prerender.routes ??= []
+          config.prerender.routes.push(...localizedEntryPages)
+        })
+      }
 
       /**
        * disable preloading/prefetching of locale files
        */
-      prepareBuildManifest(ctx, nuxt)
+      nuxt.hook('build:manifest', (manifest) => {
+        const langFiles = ctx.localeInfo
+          .flatMap(locale => locale.meta.map(m => m.path))
+          .map(x => relative(nuxt.options.srcDir, x))
+        const langPaths = [...new Set(langFiles)]
 
-      /**
-       * extend bundler
-       */
+        for (const key in manifest) {
+          if (langPaths.some(x => key.startsWith(x))) {
+            manifest[key]!.prefetch = false
+            manifest[key]!.preload = false
+          }
+        }
+      })
+
+      await setupPages(ctx, nuxt)
+
       await extendBundler(ctx, nuxt)
 
-      /**
-       * setup nitro
-       */
       await setupNitro(ctx, nuxt)
     })
   },
